commit changes pre macbook return

+++ .config/zsh/aliases.zsh
@@ -1,12 +1,18 @@
 # custom aliases
 alias diskspace="du -hxs | sort -rn"
-alias update="sudo apt-get update -y && sudo apt-get dist-upgrade -y && sudo apt autoremove -y"
+# if operating system is macOS
+if [[ "$OSTYPE" == "darwin"* ]]; then
+  alias update="brew update && brew upgrade && brew cleanup"
+  else
+    alias update="sudo apt-get update -y && sudo apt-get dist-upgrade -y && sudo apt autoremove -y"
+fi 
 alias docker="sudo docker"
 alias venvsible="source $HOME/.virtualenvs/ansible/bin/activate"
 alias py='poetry run python'
 alias clockfix='sudo ntpdate time.windows.com > /dev/null'
 alias make_pw='pwgen -y -c -s -1 -n 16 | pbcopy'
 alias remoteip='curl http://checkip.amazonaws.com/'
+alias shrug="echo '¯\\_(ツ)_/¯' | pbcopy && echo '¯\\_(ツ)_/¯'"
 
 
 # perhaps test if we are in a WSL session
@@ -17,9 +23,10 @@ if type nvim > /dev/null 2>&1; then
   alias vim="nvim"
   alias vi="nvim"
 fi
+
 # use batcat if available
-if type batcat > /dev/null 2>&1; then
-  alias cat="batcat"
+if type bat > /dev/null 2>&1; then
+  alias cat="bat" 
 fi
 
 # tmux settings
@@ -30,9 +37,14 @@ alias tmk="tmux kill-session -t"
 alias tml="tmux ls" 
 alias tmr="tmux rename-session"
 
-# list all ze things
-alias ll="ls -lha"
-alias wls="watch ls -lhat"
+# if eza is installed
+if type eza > /dev/null 2>&1; then
+  alias ls="eza --all"
+  alias ll="eza --long --all --sort=size"
+  alias tree="eza --long --tree"
+  else
+    alias ll="ls -lha"
+fi
 
 # cd navigation
 alias ..="cd .."
